loadbalancer-experimental: Narrow ewma config params from long to int (#2994)
Motivation:

We currently allow configuring a long for the parameter values that
affect the RequestTracker ewma, specifically the cancel, error, and
concurrent request penalties. There are no realistic scenarios where
a long is necessary vs an int.

Modifications:

Change the configuration parameters to an int.

+++ servicetalk-loadbalancer-experimental/src/main/java/io/servicetalk/loadbalancer/XdsHealthIndicator.java
@@ -65,8 +65,8 @@ abstract class XdsHealthIndicator<ResolvedAddress, C extends LoadBalancedConnect
     private volatile Long evictedUntilNanos;
 
     XdsHealthIndicator(final SequentialExecutor sequentialExecutor, final Executor executor,
-                       final Duration ewmaHalfLife, final long cancellationPenalty, final long errorPenalty,
-                       final long pendingRequestPenalty,
+                       final Duration ewmaHalfLife, final int cancellationPenalty, final int errorPenalty,
+                       final int pendingRequestPenalty,
                        final boolean cancellationIsError, final ResolvedAddress address, String lbDescription,
                        final HostObserver hostObserver) {
         super(requireNonNull(ewmaHalfLife, "ewmaHalfLife").toNanos(),
